mylinkedin/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
jinja2>=3.1.4
itsdangerous>=2.2.0
starlette>=0.41.0
requests>=2.32.0
python-dotenv>=1.0.0
